enhance(Footer): add Footer interfaces, display Footer content by locale
