+++ src/components/Footer/Footer.tsx
@@ -1,42 +1,19 @@
 import React from 'react'
-import Image from 'next/image'
 
 import { LogoWithLabelBright } from '@/assets'
-import { Facebook, Github, Instagram, LinkedIn, Youtube } from './assets'
 
 import * as S from './Footer.styles'
 
-const Footer = () => {
-  const socialMediaMap = [
-    {
-      icon: Instagram,
-      link: 'https://www.instagram.com/robocinufpe/',
-    },
-    {
-      icon: Github,
-      link: 'https://github.com/robocin',
-    },
-    {
-      icon: Facebook,
-      link: 'https://www.facebook.com/robocin',
-    },
-    {
-      icon: LinkedIn,
-      link: 'https://www.linkedin.com/company/robocin',
-    },
-    {
-      icon: Youtube,
-      link: 'https://www.youtube.com/@robocinufpe',
-    },
-  ]
+import { FooterProps } from './interfaces'
 
+const Footer = ({ translate, socialMediaMap }: FooterProps) => {
   return (
     <S.Container>
       <S.Content>
-        <h3>Nosso contato</h3>
+        <h3>{translate?.contact}</h3>
         <p>robocin@cin.ufpe.br</p>
         <S.SocialMediaWrapper>
-          {socialMediaMap.map((socialMedia, index) => (
+          {socialMediaMap?.map((socialMedia, index) => (
             <a
               href={socialMedia.link}
               target="_blank"
@@ -50,11 +27,8 @@ const Footer = () => {
       </S.Content>
       <S.Logo src={LogoWithLabelBright} alt="Logo" priority />
       <S.Content align="right">
-        <h3>Onde estamos</h3>
-        <p>
-          Av. Jorn. Aníbal Fernandes, s/n - Cidade Universitária, Recife - PE,
-          50740-560
-        </p>
+        <h3>{translate?.location.label}</h3>
+        <p>{translate?.location.address}</p>
       </S.Content>
     </S.Container>
   )
